Data Structures. Ruby. BinaryTree. Reorganizes namespacing. Deletes BinaryTree module. Creates BinaryTreeNode class instead.
