--- data-structures/ruby/binary_tree/binary_tree_node.rb
@@ -1,12 +1,0 @@
-module BinaryTree
-
-  def BinaryTree(value, left = nil, right = nil)
-    case value
-    when BinaryTree::Node
-      value
-    else
-      BinaryTree::Node.new(value, left, right)
-    end
-  end
-
-end
